requirements.txt
Flask==2.3.2
google-cloud-aiplatform==1.26.0